refactor: rename dataproxycleaner to dataproxyrefresher
