+++ .doc/secrets.md
@@ -73,9 +73,9 @@ kubectl \
   --from-literal="MONGO_USERNAME=data-proxy-admin" \
   --from-literal="MONGO_PASSWORD=password"
   
-# dataproxycleaner connects to DataProxy MongoDB.
+# dataproxyrefresher connects to DataProxy MongoDB.
 kubectl \
-  create secret generic dataproxycleaner-secret \
+  create secret generic dataproxyrefresher-secret \
   --from-literal="MONGO_USERNAME=data-proxy-admin" \
   --from-literal="MONGO_PASSWORD=password"
 ```
