--- services/cronjob/dataproxycleaner/Makefile
@@ -1,3 +1,0 @@
-docker-build:
-	cd ../../../ && \
-	docker build -f services/cronjob/dataproxycleaner/docker/Dockerfile-prod -t murmurations/dataproxycleaner .
