--- services/cronjob/dataproxycleaner/docker/Dockerfile-prod
@@ -1,12 +1,0 @@
-FROM golang:1.17.5-alpine as build
-
-WORKDIR /src/dataproxycleaner
-ADD . /src/dataproxycleaner
-RUN CGO_ENABLED=0 go build -o /bin/dataproxycleaner ./services/cronjob/dataproxycleaner/cmd/dataproxycleaner
-
-FROM ubuntu:20.04
-
-RUN apt-get update && apt-get install -y --no-install-recommends ca-certificates
-COPY --from=build /bin/dataproxycleaner /app/dataproxycleaner
-EXPOSE 8000
-CMD ["/app/dataproxycleaner"]
