+++ services/cronjob/dataproxyrefresher/Makefile
@@ -1,0 +1,3 @@
+docker-build:
+	cd ../../../ && \
+	docker build -f services/cronjob/dataproxyrefresher/docker/Dockerfile-prod -t murmurations/dataproxyrefresher .
