+++ services/cronjob/dataproxyrefresher/docker/Dockerfile-prod
@@ -1,0 +1,12 @@
+FROM golang:1.17.5-alpine as build
+
+WORKDIR /src/dataproxyrefresher
+ADD . /src/dataproxyrefresher
+RUN CGO_ENABLED=0 go build -o /bin/dataproxyrefresher ./services/cronjob/dataproxyrefresher/cmd/dataproxyrefresher
+
+FROM ubuntu:20.04
+
+RUN apt-get update && apt-get install -y --no-install-recommends ca-certificates
+COPY --from=build /bin/dataproxyrefresher /app/dataproxyrefresher
+EXPOSE 8000
+CMD ["/app/dataproxyrefresher"]
