+++ skaffold.yaml
@@ -40,10 +40,10 @@ build:
       context: .
       docker:
         dockerfile: services/cronjob/dataproxyupdater/docker/Dockerfile-dev
-    - image: murmurations/dataproxycleaner
+    - image: murmurations/dataproxyrefresher
       context: .
       docker:
-        dockerfile: services/cronjob/dataproxycleaner/docker/Dockerfile-dev
+        dockerfile: services/cronjob/dataproxyrefresher/docker/Dockerfile-dev
 deploy:
   helm:
     releases:
@@ -59,7 +59,7 @@ deploy:
         geoip.image: murmurations/geoip
         dataproxy.image: murmurations/dataproxy
         dataproxyupdater.image: murmurations/dataproxyupdater
-        dataproxycleaner.image: murmurations/dataproxycleaner
+        dataproxyrefresher.image: murmurations/dataproxyrefresher
       # Skaffold will try to parse the value part of each key-value pair using environment variables in the system
       setValueTemplates:
         global.sourcepath: "{{.SOURCEPATH}}"
